update landing page
