+++ gatsby-config.js
@@ -1,7 +1,5 @@
 require('dotenv').config()
 
-const googleTrackingCode = (process.env.NODE_ENV === `production`) ? "UA-113726758-1" : "UA-113726758-2"
-
 module.exports = {
   siteMetadata: {
     title: 'Gatsby Default Starter',
