+++ package.json
@@ -27,7 +27,7 @@
     "react-iframe": "^1.2.0",
     "semantic-ui-css": "^2.3.1",
     "semantic-ui-react": "^0.81.1",
-    "styled-components": "^3.3.2"
+    "styled-components": "^4.1.1"
   },
   "keywords": [
     "gatsby"
